moves survivor form to its own component
