+++ lib/types/survivor.ts
@@ -1,0 +1,22 @@
+export type Survivor = {
+  id: number;
+  settlement: number;
+  name: string;
+  born: number;
+  gender: "M" | "F";
+  status: "alive" | "dead" | "retired";
+  huntXp: number;
+  survival: number;
+  movement: number;
+  accuracy: number;
+  strength: number;
+  evasion: number;
+  luck: number;
+  speed: number;
+  insanity: number;
+  systemicPressure: number;
+  torment: number;
+  lumi: number;
+  courage: number;
+  understanding: number;
+};
